implement evaluateFiles for CWL
