+++ cwl/src/main/scala/cwl/CwlExpressionCommandPart.scala
@@ -1,6 +1,5 @@
 package cwl
 
-import cwl.CwlWomExpression._
 import wdl.values.WdlValue
 import wom.CommandPart
 import wom.expression.IoFunctionSet
