+++ cwl/src/main/scala/cwl/ParameterContext.scala
@@ -2,6 +2,7 @@ package cwl
 
 import wdl.types._
 import wdl.values._
+import wom.expression.IoFunctionSet
 
 object ParameterContext {
   val Empty = ParameterContext(
@@ -11,4 +12,15 @@ object ParameterContext {
   )
 }
 
-case class ParameterContext(inputs: WdlValue, self: WdlValue, runtime: WdlValue)
+case class ParameterContext(inputs: WdlValue, self: WdlValue, runtime: WdlValue) {
+  def withInputs(inputValues: Map[String, WdlValue], ioFunctionSet: IoFunctionSet): ParameterContext = {
+    val wdlValueType = inputValues.values.headOption.map(_.wdlType).getOrElse(WdlNothingType)
+    copy(
+      inputs = WdlMap(
+        WdlMapType(WdlStringType, wdlValueType),
+        // TODO: WOM: convert inputValues (including WdlFile?) to inputs using the ioFunctionSet
+        inputValues map { case (name, wdlValue) => WdlString(name) -> wdlValue }
+      )
+    )
+  }
+}
